Conflicts:
	src/demos/trackVehicle/demo_trackVehicle.cpp
	src/demos/trackVehicle/subsys/ChTrackVehicle.cpp
	src/demos/trackVehicle/subsys/suspension/TorsionArmSuspension.cpp
	src/demos/trackVehicle/subsys/suspension/TorsionArmSuspension.h
	src/demos/trackVehicle/subsys/trackSystem/TrackSystem.cpp
	src/demos/trackVehicle/subsys/trackSystem/TrackSystem.h
	src/demos/trackVehicle/subsys/trackSystem/TrackSystemM113.cpp
	src/demos/trackVehicle/subsys/trackVehicle/DriveChain.cpp
	src/demos/trackVehicle/subsys/trackVehicle/TrackVehicle.cpp
	src/demos/trackVehicle/test_driveChain.cpp

+++ src/chrono/utils/ChFilters.cpp
@@ -2,7 +2,7 @@
 // PROJECT CHRONO - http://projectchrono.org
 //
 // Copyright (c) 2014 projectchrono.org
-// All right reserved.
+// All rights reserved.
 //
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file at the top level of the distribution and at
@@ -34,7 +34,7 @@ double ChRunningAverage::Add(double val) {
 
 // -----------------------------------------------------------------------------
 ChMovingAverage::ChMovingAverage(const std::valarray<double>& data, int n) : m_n(n) {
-    int np = data.size();
+    int np = (int) data.size();
     m_out.resize(np);
 
     // Start and end of data
